更新MultiTimer 软件定时器无限扩展

+++ RNA_Multi_Timer_SysTick/sys/delay.h
@@ -23,7 +23,7 @@
 #include <ti/devices/msp432p4xx/driverlib/driverlib.h>
 
 void delay_init(void);
-void delay_ms(uint32_t nms);
+void delay_ms(uint16_t nms);
 void delay_us(uint32_t nus);
 
 #endif
